+++ projects/Mallard/src/components/layout/slide-card/helpers.ts
@@ -1,0 +1,1 @@
+export const dismissAt = 100
