+++ projects/Mallard/src/components/layout/slide-card/index.tsx
@@ -1,0 +1,111 @@
+import React, { useState, useEffect, ReactNode, useRef } from 'react'
+import { Animated, Platform, StyleSheet } from 'react-native'
+import { Header } from './header'
+import { dismissAt } from './helpers'
+
+/* 
+This is the swipey contraption that contains an article.
+*/
+
+const notchInsetSize = Platform.OS === 'ios' ? 50 : 0
+const styles = StyleSheet.create({
+    container: {
+        marginTop: notchInsetSize,
+        shadowColor: '#000',
+        shadowOffset: {
+            width: 0,
+            height: -2,
+        },
+        shadowOpacity: 0.25,
+        shadowRadius: 3.84,
+        elevation: 5,
+        flex: 1,
+    },
+})
+
+export const SlideCard = ({
+    children,
+    headerStyle,
+    backgroundColor,
+    onDismiss,
+}: {
+    children: ReactNode
+    headerStyle: {}
+    backgroundColor: string | undefined
+    onDismiss: () => void
+}) => {
+    const [scrollY] = useState(() => new Animated.Value(1))
+    const [cardOffset] = useState(() => new Animated.Value(0))
+    useEffect(() => {
+        scrollY.addListener(({ value }) => {
+            if (value < dismissAt * -1) {
+                onDismiss()
+            }
+        })
+    }, [])
+    return (
+        <Animated.View
+            style={[
+                styles.container,
+                {
+                    transform: [
+                        {
+                            translateY: Animated.add(
+                                scrollY.interpolate({
+                                    inputRange: [dismissAt * -1, 0],
+                                    outputRange: [dismissAt, 0],
+                                    extrapolate: 'clamp',
+                                }),
+                                cardOffset,
+                            ),
+                        },
+                    ],
+                },
+            ]}
+        >
+            <Header
+                style={headerStyle}
+                {...{
+                    scrollY,
+                    cardOffset,
+                    onDismiss,
+                }}
+            />
+            <Animated.ScrollView
+                scrollEventThrottle={1}
+                contentContainerStyle={{
+                    flexGrow: 1,
+                }}
+                style={{ backgroundColor }}
+                onScroll={Animated.event(
+                    [
+                        {
+                            nativeEvent: {
+                                contentOffset: {
+                                    y: scrollY,
+                                },
+                            },
+                        },
+                    ],
+                    { useNativeDriver: true },
+                )}
+            >
+                <Animated.View
+                    style={{
+                        transform: [
+                            {
+                                translateY: scrollY.interpolate({
+                                    inputRange: [-100, 0],
+                                    outputRange: [-100, 0],
+                                    extrapolate: 'clamp',
+                                }),
+                            },
+                        ],
+                    }}
+                >
+                    {children}
+                </Animated.View>
+            </Animated.ScrollView>
+        </Animated.View>
+    )
+}
